Added Creature card type
Note about comments, any comments starting and ending with // are added
to be temporary comments to warn about potential bugs or explain a new
way of doing something. Feel free to delete them if you feel the comment
is not needed anymore. The same goes for the controls listed in the
ADDED FOR TESTING PURPOSES comment at the top of Form1.vb.